+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- [#23](https://github.com/zendframework/zend-expressive-helpers/pull/23) adds support to UrlHelper for generating
+  default route w/o parameters
 
 ### Deprecated